Move AlertsDate & add tests for it
Starts tests by:
- creating a ./tests folder (as a package)
- moving AlertsDate to ./lib/utils.py so it can be
  imported
- creating a test for AlertsDate
