+++ build.py
@@ -11,6 +11,7 @@ from jinja2 import (
     PrefixLoader,
 )
 from notifications_utils.formatters import formatted_list
+from lib.utils import AlertsDate
 
 repo = Path('.')
 src = repo / 'src'
@@ -25,25 +26,6 @@ jinja_loader = ChoiceLoader([
         'govuk_frontend_jinja': PackageLoader('govuk_frontend_jinja')
     })
 ])
-
-
-class AlertsDate(object):
-    """Makes a datetime available in different formats"""
-
-    def __init__(self, _datetime):
-        self._datetime = _datetime
-
-    @property
-    def as_lang(self, lang='en-GB'):
-        return '{dt.day} {dt:%B} {dt:%Y} at {dt:%H}:{dt:%M}'.format(dt=self._datetime)
-
-    @property
-    def as_iso8601(self):
-        return self._datetime.isoformat()
-
-    @property
-    def as_datetime(self):
-        return self._datetime
 
 
 def file_fingerprint(path):
